# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Small Town Watchdog API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /multi-year-timeline:
    get:
      operationId: getMultiYearTimeline
      tags: [compare]
      summary: Get year-by-year document timeline for an entity
      parameters:
        - name: entityId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiYearTimeline"
        "404":
          description: Not found
  /compare:
    get:
      operationId: compareYears
      tags: [compare]
      summary: Compare two years of data for an entity
      parameters:
        - name: entityId
          in: query
          required: true
          schema:
            type: integer
        - name: year1
          in: query
          required: true
          schema:
            type: integer
        - name: year2
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YearComparison"
...
  /ballot-items/{id}:
    patch:
      operationId: updateBallotItem
      tags: [elections]
      summary: Update a ballot item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BallotItemPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BallotItem"
        "404":
          description: Not found
    delete:
      operationId: deleteBallotItem
      tags: [elections]
      summary: Delete a ballot item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
        "404":
          description: Not found
  /ballot-items/{id}/explain:
    post:
      operationId: explainBallotItem
      tags: [elections]
      summary: AI-generate a neutral plain-English explanation of a ballot item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BallotItem"
        "404":
          description: Not found
...
  /civic-items:
    get:
      operationId: listCivicItems
      tags: [civic-items]
      summary: List approved civic items for a location
      parameters:
        - name: locationId
          in: query
          schema:
            type: integer
        - name: stateCode
          in: query
          schema:
            type: string
        - name: countyParish
          in: query
          schema:
            type: string
        - name: itemType
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CivicItem"
  /admin/civic-items:
    get:
      operationId: listAdminCivicItems
      tags: [civic-items]
      summary: Admin — list all civic items regardless of review status
      parameters:
        - name: locationId
          in: query
          schema:
            type: integer
        - name: stateCode
          in: query
          schema:
            type: string
        - name: countyParish
          in: query
          schema:
            type: string
        - name: adminReviewStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CivicItem"
  /admin/civic-items/{id}:
    patch:
      operationId: updateCivicItemReview
      tags: [civic-items]
      summary: Admin — update review status of a civic item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CivicItemPatch"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CivicItem"
        "404":
          description: Not found
...
components:
  schemas:
...
    YearComparison:
      type: object
      required: [entityId, entityName, year1, year2, changes, aiSummary]
      properties:
        entityId:
          type: integer
        entityName:
          type: string
        year1:
          type: integer
        year2:
          type: integer
        changes:
          type: array
          items:
            $ref: "#/components/schemas/YearChange"
        aiSummary:
          type: string
        year1DocumentCount:
          type: integer
        year2DocumentCount:
          type: integer
...
    MultiYearTimeline:
      type: object
      required: [entityId, entityName, years]
      properties:
        entityId:
          type: integer
        entityName:
          type: string
        years:
          type: array
          items:
            $ref: "#/components/schemas/YearSummary"
    YearSummary:
      type: object
      required: [year, documentCount, hasData, brokenSourceCount, missingSourceCount, documents]
      properties:
        year:
          type: integer
        documentCount:
          type: integer
        hasData:
          type: boolean
        totalAmount:
          type: ["number", "null"]
        brokenSourceCount:
          type: integer
        missingSourceCount:
          type: integer
        documents:
          type: array
          items:
            type: object
            required: [id, title, docType, sourceStatus]
            properties:
              id:
                type: integer
              title:
                type: string
              docType:
                type: string
              amountInvolved:
                type: ["number", "null"]
              sourceStatus:
                type: string
              sourceUrl:
                type: ["string", "null"]
    YearChange:
      type: object
      required: [category, label, year1Value, year2Value, changePercent, redFlagLevel]
      properties:
        category:
          type: string
        label:
          type: string
        year1Value:
          type: ["number", "null"]
        year2Value:
          type: ["number", "null"]
        changePercent:
          type: ["number", "null"]
        changeDescription:
          type: ["string", "null"]
        redFlagLevel:
          type: string
...
    BallotItem:
      type: object
      required: [id, electionId, title, itemType, isAiGenerated, createdAt]
      properties:
        id:
          type: integer
        electionId:
          type: integer
        title:
          type: string
        itemType:
          type: string
        description:
          type: ["string", "null"]
        officialText:
          type: ["string", "null"]
        yesMeans:
          type: ["string", "null"]
        noMeans:
          type: ["string", "null"]
        whoPays:
          type: ["string", "null"]
        amountInvolved:
          type: ["string", "null"]
        duration:
          type: ["string", "null"]
        receivingBody:
          type: ["string", "null"]
        changeType:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        isAiGenerated:
          type: boolean
        createdAt:
          type: string
...
    BallotItemPatch:
      type: object
      properties:
        title:
          type: string
        itemType:
          type: string
        description:
          type: string
        officialText:
          type: string
        yesMeans:
          type: string
        noMeans:
          type: string
        whoPays:
          type: string
        amountInvolved:
          type: string
        duration:
          type: string
        receivingBody:
          type: string
        changeType:
          type: string
        sourceUrl:
          type: string
...
    CivicItem:
      type: object
      required: [id, itemType, title, redFlagLevel, sourceStatus, adminReviewStatus, createdAt, updatedAt]
      properties:
        id:
          type: integer
        locationId:
          type: ["integer", "null"]
        entityId:
          type: ["integer", "null"]
        stateCode:
          type: ["string", "null"]
        countyParish:
          type: ["string", "null"]
        scope:
          type: string
          enum: [statewide, county_parish, city, entity]
        itemType:
          type: string
        title:
          type: string
        eventDate:
          type: ["string", "null"]
        meetingDate:
          type: ["string", "null"]
        electionDate:
          type: ["string", "null"]
        deadlineDate:
          type: ["string", "null"]
        amountInvolved:
          type: ["string", "null"]
        voteResult:
          type: ["string", "null"]
        sourceTitle:
          type: ["string", "null"]
        sourceAgency:
          type: ["string", "null"]
        sourceUrl:
          type: ["string", "null"]
        sourceDate:
          type: ["string", "null"]
        originalText:
          type: ["string", "null"]
        aiSummary:
          type: ["string", "null"]
        aiSummaryNotice:
          type: ["string", "null"]
        redFlagLevel:
          type: string
        sourceStatus:
          type: string
        adminReviewStatus:
          type: string
        lastVerifiedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    CivicItemPatch:
      type: object
      properties:
        adminReviewStatus:
          type: string
...